requirements.txt
transformers
torch 
torchvision 
torchaudio
pillow 
requests
qwen-vl-utils
accelerate
bitsandbytes